feat: formatPrice 함수 useFormatPrice 공유 hooks로 분리
