+++ src/basic/shared/hooks/index.ts
@@ -7,3 +7,6 @@ export { useLocalStorage } from "./useLocalStorage";
 
 // Search hooks (범용적으로 사용 가능)
 export { useSearch } from "./useSearch";
+
+// Format hooks (UI 관련 범용 훅)
+export { useFormatPrice } from "./useFormatPrice";
